18-02-2023

+++ database/migrations/2023_02_18_080333_create_pr_actual_vs_schedules_table.php
@@ -19,9 +19,9 @@ class CreatePrActualVsSchedulesTable extends Migration
             $table->bigInteger('pr_detail_id')->unsigned();
             $table->bigInteger('pr_contractor_id')->unsigned();
             $table->date('month');
-            $table->tinyInteger('schedule_progress')->unsigned();
-            $table->tinyInteger('actual_progress')->unsigned()->nullable();
-            $table->tinyInteger('current_month_progress')->unsigned()->nullable();
+            $table->float('schedule_progress')->unsigned();
+            $table->float('actual_progress')->unsigned()->nullable();
+            $table->float('current_month_progress')->unsigned()->nullable();
             $table->timestamps();
             $table->foreign('pr_detail_id')->references('id')->on('pr_details')->onDelete('cascade');
             $table->foreign('pr_contractor_id')->references('id')->on('pr_contractors')->onDelete('cascade');
